Fix to avoid SimpleMbedCloudClient from re-initialising filesystem storage
As of MCC 2.1.0, arm_uc_blockdevice is no longer referenced from MCC, and is therefore pointless to keep in SMCC.
As such, SMCC only tracks the blockdevice backing the filesystem now, not the one backing both filesystem and upgrade storage. That solves the issue of the internal remount logic remounting the filesystem onto the full blockdevice, instead of the potentially smaller slicingblockdevice having been assigned to the filesystem earlier.